# src/main/resources/OpenApi 3.yaml
openapi: 3.0.0
info:
  title: Gestor Info API
  version: 1.0.0
  description: API para gestionar información
servers:
  - url: 'https://api.gestor.com'
...
paths:
  /users:
    get:
      tags:
        - users
      summary: Obtener todos los usuarios
      description: |
        Esta operación permite obtener una lista de todos los usuarios registrados.
        -
        -
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: Devuelve exitosamente una lista de todos los usuarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InternalServerError'

    post:
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Usuario creado correctamente
  /users/{id}:
    get:
      summary: Obtener un usuario por ID
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Éxito
        '404':
          description: Usuario no encontrado

    put:
      summary: Actualizar un usuario por ID
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Usuario actualizado correctamente
        '404':
          description: Usuario no encontrado

    delete:
      summary: Eliminar un usuario por ID
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Usuario eliminado correctamente
        '404':
          description: Usuario no encontrado
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
        fullname:
          type: string
        email:
          type: string
      required:
        - username
        - password
        - fullname
        - email
...
  parameters:
    userId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...